feat(point) : DIVIDE - Point class
